#42 fixed HomePage according to review comments
